some work for version 2

+++ fltk/src/window.rs
@@ -1030,7 +1030,7 @@ impl AndroidWindow {
     pub fn default() -> Self {
         let (w, h) = screen_size();
         let mut w = AndroidWindow {
-            win: Window::new(0, 30, w as i32, h as i32 - 30, ""),
+            win: Window::new(0, 30, w, h - 30, ""),
         };
         w.win.set_color(Color::White);
         w
